src/requirements.txt
gradio==6.6.0
pillow
numpy
pennylane
opencv-python
pennylane-lightning
qiskit
matplotlib